profile updated done

+++ client/src/Redux/Actions/profile.js
@@ -1,5 +1,13 @@
 import axios from "axios";
-import { GET_PROFILE, PROFILE_ERROR, UPDATE_PROFILE, ACCOUNT_DELETED, CLEAR_PROFILE } from "./types";
+import {
+  GET_PROFILE,
+  PROFILE_ERROR,
+  UPDATE_PROFILE,
+  ACCOUNT_DELETED,
+  CLEAR_PROFILE,
+  GET_PROFILES,
+  GET_REPOS
+} from "./types";
 import { setAlert } from "./alert";
 
 //get current profile
@@ -9,6 +17,58 @@ export const getCurrentProfile = () => async dispatch => {
 
     dispatch({
       type: GET_PROFILE,
+      payload: res.data
+    });
+  } catch (err) {
+    dispatch({
+      type: PROFILE_ERROR,
+      payload: { msg: err.response.statusText, status: err.response.status }
+    });
+  }
+};
+
+//get all profiles
+export const getProfiles = () => async dispatch => {
+  dispatch({ type: CLEAR_PROFILE });
+  try {
+    const res = await axios.get("/api/profile");
+
+    dispatch({
+      type: GET_PROFILES,
+      payload: res.data
+    });
+  } catch (err) {
+    dispatch({
+      type: PROFILE_ERROR,
+      payload: { msg: err.response.statusText, status: err.response.status }
+    });
+  }
+};
+
+//get profile by Id
+export const getProfileById = userId => async dispatch => {
+  try {
+    const res = await axios.get(`/api/profile/user/${userId}`);
+
+    dispatch({
+      type: GET_PROFILE,
+      payload: res.data
+    });
+  } catch (err) {
+    dispatch({
+      type: PROFILE_ERROR,
+      payload: { msg: err.response.statusText, status: err.response.status }
+    });
+  }
+};
+
+//get github repos
+export const getGithubRepos = username => async dispatch => {
+  try {
+    const res = await axios.get(`/api/profile/github/${username}`);
+
+    dispatch({
+      type: GET_REPOS,
       payload: res.data
     });
   } catch (err) {
@@ -65,7 +125,7 @@ export const addExperience = (formData, history) => async dispatch => {
         "Content-Type": "application/json"
       }
     };
-    const res = await axios.post("/api/profile/experience", formData, config);
+    const res = await axios.put("/api/profile/experience", formData, config);
 
     dispatch({
       type: UPDATE_PROFILE,
@@ -96,7 +156,7 @@ export const addEducation = (formData, history) => async dispatch => {
         "Content-Type": "application/json"
       }
     };
-    const res = await axios.post("/api/profile/education", formData, config);
+    const res = await axios.put("/api/profile/education", formData, config);
 
     dispatch({
       type: UPDATE_PROFILE,
@@ -154,20 +214,21 @@ export const deleteEducation = id => async dispatch => {
   }
 };
 
+//Delete Account & profile
+export const deleteAccount = () => async dispatch => {
+  if (window.confirm("Are You Sure? This cannot be undone")) {
+    try {
+      await axios.delete("/api/profile");
 
-//Delete Account
-export const deleteAccount = ()> async dispatch => {
-  try {
-    const res = await axios.delete('/api/profile');
+      dispatch({ type: CLEAR_PROFILE });
+      dispatch({ type: ACCOUNT_DELETED });
 
-    dispatch({ type: CLEAR_PROFILE });
-    dispatch({ type: ACCOUNT_DELETED });
-
-    dispatch(setAlert('Your Acccount Permanently Deleted'));
-  } catch (err) {
-    dispatch({
-      type: PROFILE_ERROR,
-      payload: { msg: err.response.statusText, status: err.response.status }
-    });
+      dispatch(setAlert("Your Acccount Permanently Deleted"));
+    } catch (err) {
+      dispatch({
+        type: PROFILE_ERROR,
+        payload: { msg: err.response.statusText, status: err.response.status }
+      });
+    }
   }
 };
